+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+-   Numerous updates to `JWTPayload` class
+
 ## 3.0.9 (2022-08-16)
 
 ### Added